@@ -5,6 +5,6 @@ set(SRCS
 )
 
 # glslang requires C++11 at a minimum to compile.
-add_definitions(-std=c++11)
+add_definitions(-std=c++14)
 
 add_library(SlippiLib STATIC ${SRCS})
