@@ -183,6 +183,7 @@ void SConfig::SaveInterfaceSettings(IniFile& ini)
 	interface->Set("LanguageCode", m_InterfaceLanguage);
 	interface->Set("ShowToolbar", m_InterfaceToolbar);
 	interface->Set("ShowStatusbar", m_InterfaceStatusbar);
+	interface->Set("ShowSeekbar", m_InterfaceSeekbar);
 	interface->Set("ShowLogWindow", m_InterfaceLogWindow);
 	interface->Set("ShowLogConfigWindow", m_InterfaceLogConfigWindow);
 	interface->Set("ExtendedFPSInfo", m_InterfaceExtendedFPSInfo);
@@ -266,6 +267,7 @@ void SConfig::SaveCoreSettings(IniFile& ini)
 	core->Set("FPRF", bFPRF);
 	core->Set("AccurateNaNs", bAccurateNaNs);
 	core->Set("DefaultISO", m_strDefaultISO);
+	core->Set("BootDefaultISO", bBootDefaultISO);
 	core->Set("DVDRoot", m_strDVDRoot);
 	core->Set("Apploader", m_strApploader);
 	core->Set("EnableCheats", bEnableCheats);
@@ -478,26 +480,32 @@ void SConfig::LoadGeneralSettings(IniFile& ini)
 	general->Get("DumpPath", &m_DumpPath);
 	CreateDumpPath(m_DumpPath);
 	general->Get("WirelessMac", &m_WirelessMac);
-	general->Get("WiiSDCardPath", &m_strWiiSDCardPath, File::GetUserPath(F_WIISDCARD_IDX));
+	general->Get("WiiSDCardPath", &m_strWiiSDCardPath);
 	File::SetUserPath(F_WIISDCARD_IDX, m_strWiiSDCardPath);
 }
 
 void SConfig::LoadInterfaceSettings(IniFile& ini)
 {
 	IniFile::Section* interface = ini.GetOrCreateSection("Interface");
-
-	interface->Get("ConfirmStop", &bConfirmStop, true);
+#ifdef IS_PLAYBACK
+	interface->Get("UsePanicHandlers", &bUsePanicHandlers, false);
+	interface->Get("OnScreenDisplayMessages", &bOnScreenDisplayMessages, false);
+#else
 	interface->Get("UsePanicHandlers", &bUsePanicHandlers, true);
 	interface->Get("OnScreenDisplayMessages", &bOnScreenDisplayMessages, true);
-	interface->Get("HideCursor", &bHideCursor, false);
+	
+#endif
+	interface->Get("HideCursor", &bHideCursor, true);
+	interface->Get("ConfirmStop", &bConfirmStop, false);
 	interface->Get("AutoHideCursor", &bAutoHideCursor, false);
 	interface->Get("MainWindowPosX", &iPosX, INT_MIN);
 	interface->Get("MainWindowPosY", &iPosY, INT_MIN);
-	interface->Get("MainWindowWidth", &iWidth, -1);
-	interface->Get("MainWindowHeight", &iHeight, -1);
+	interface->Get("MainWindowWidth", &iWidth, 640);
+	interface->Get("MainWindowHeight", &iHeight, 430);
 	interface->Get("LanguageCode", &m_InterfaceLanguage, "");
 	interface->Get("ShowToolbar", &m_InterfaceToolbar, true);
 	interface->Get("ShowStatusbar", &m_InterfaceStatusbar, true);
+	interface->Get("ShowSeekbar", &m_InterfaceSeekbar, true);
 	interface->Get("ShowLogWindow", &m_InterfaceLogWindow, false);
 	interface->Get("ShowLogConfigWindow", &m_InterfaceLogConfigWindow, false);
 	interface->Get("ExtendedFPSInfo", &m_InterfaceExtendedFPSInfo, false);
@@ -512,7 +520,11 @@ void SConfig::LoadDisplaySettings(IniFile& ini)
 
 	display->Get("Fullscreen", &bFullscreen, false);
 	display->Get("FullscreenResolution", &strFullscreenResolution, "Auto");
+#ifdef IS_PLAYBACK
+	display->Get("RenderToMain", &bRenderToMain, true);
+#else
 	display->Get("RenderToMain", &bRenderToMain, false);
+#endif
 	display->Get("RenderWindowXPos", &iRenderWindowXPos, -1);
 	display->Get("RenderWindowYPos", &iRenderWindowYPos, -1);
 	display->Get("RenderWindowWidth", &iRenderWindowWidth, 640);
@@ -559,11 +571,11 @@ void SConfig::LoadGameListSettings(IniFile& ini)
 	gamelist->Get("ColumnPlatform", &m_showSystemColumn, true);
 	gamelist->Get("ColumnBanner", &m_showBannerColumn, true);
 	gamelist->Get("ColumnNotes", &m_showMakerColumn, true);
-	gamelist->Get("ColumnFileName", &m_showFileNameColumn, false);
-	gamelist->Get("ColumnID", &m_showIDColumn, false);
+	gamelist->Get("ColumnFileName", &m_showFileNameColumn, true);
+	gamelist->Get("ColumnID", &m_showIDColumn, true);
 	gamelist->Get("ColumnRegion", &m_showRegionColumn, true);
 	gamelist->Get("ColumnSize", &m_showSizeColumn, true);
-	gamelist->Get("ColumnState", &m_showStateColumn, true);
+	gamelist->Get("ColumnState", &m_showStateColumn, false);
 }
 
 void SConfig::LoadCoreSettings(IniFile& ini)
@@ -580,43 +592,44 @@ void SConfig::LoadCoreSettings(IniFile& ini)
 #endif
 	core->Get("Fastmem", &bFastmem, true);
 	core->Get("DSPHLE", &bDSPHLE, true);
-	core->Get("TimingVariance", &iTimingVariance, 40);
+	core->Get("TimingVariance", &iTimingVariance, 8);
+#ifdef IS_PLAYBACK
+	core->Get("CPUThread", &bCPUThread, false);
+#else
 	core->Get("CPUThread", &bCPUThread, true);
+#endif
 	core->Get("SyncOnSkipIdle", &bSyncGPUOnSkipIdleHack, true);
 	core->Get("DefaultISO", &m_strDefaultISO);
+	core->Get("BootDefaultISO", &bBootDefaultISO, false);
 	core->Get("DVDRoot", &m_strDVDRoot);
 	core->Get("Apploader", &m_strApploader);
-	core->Get("EnableCheats", &bEnableCheats, false);
+	core->Get("EnableCheats", &bEnableCheats, true);
 	core->Get("SelectedLanguage", &SelectedLanguage, 0);
 	core->Get("OverrideGCLang", &bOverrideGCLanguage, false);
 	core->Get("DPL2Decoder", &bDPL2Decoder, false);
 	core->Get("TimeStretching", &bTimeStretching, false);
 	core->Get("RSHACK", &bRSHACK, false);
-	core->Get("Latency", &iLatency, 2);
+	core->Get("Latency", &iLatency, 0);
 	core->Get("SlippiOnlineDelay", &m_slippiOnlineDelay, 2);
 	core->Get("SlippiSaveReplays", &m_slippiSaveReplays, true);
 	core->Get("SlippiReplayMonthFolders", &m_slippiReplayMonthFolders, false);
-#ifdef _WIN32
-	core->Get("SlippiReplayDir", &m_strSlippiReplayDir,
-		File::GetHomeDirectory() + "\\Slippi");
-#else
-	core->Get("SlippiReplayDir", &m_strSlippiReplayDir,
-		File::GetHomeDirectory() + DIR_SEP + "Slippi");
-#endif
+	std::string default_replay_dir = File::GetHomeDirectory() + DIR_SEP + "Slippi";
+	core->Get("SlippiReplayDir", &m_strSlippiReplayDir, default_replay_dir);
+	if (m_strSlippiReplayDir.empty())
+		m_strSlippiReplayDir = default_replay_dir;
 	core->Get("MemcardAPath", &m_strMemoryCardA);
 	core->Get("MemcardBPath", &m_strMemoryCardB);
 	core->Get("AgpCartAPath", &m_strGbaCartA);
 	core->Get("AgpCartBPath", &m_strGbaCartB);
-	core->Get("SlotA", (int*)&m_EXIDevice[0], EXIDEVICE_MEMORYCARD);
-	core->Get("SlotB", (int*)&m_EXIDevice[1], EXIDEVICE_NONE);
+	core->Get("SlotA", (int *)&m_EXIDevice[0], EXIDEVICE_NONE);
+	core->Get("SlotB", (int*)&m_EXIDevice[1], EXIDEVICE_SLIPPI);
 	core->Get("SerialPort1", (int*)&m_EXIDevice[2], EXIDEVICE_NONE);
 	core->Get("BBA_MAC", &m_bba_mac);
 	core->Get("TimeProfiling", &bJITILTimeProfiling, false);
 	core->Get("OutputIR", &bJITILOutputIR, false);
 	for (int i = 0; i < MAX_SI_CHANNELS; ++i)
 	{
-		core->Get(StringFromFormat("SIDevice%i", i), (u32*)&m_SIDevice[i],
-			(i == 0) ? SIDEVICE_GC_CONTROLLER : SIDEVICE_NONE);
+		core->Get(StringFromFormat("SIDevice%i", i), (u32*)&m_SIDevice[i], SIDEVICE_WIIU_ADAPTER);
 		core->Get(StringFromFormat("AdapterRumble%i", i), &m_AdapterRumble[i], true);
 		core->Get(StringFromFormat("SimulateKonga%i", i), &m_AdapterKonga[i], false);
 	}
@@ -677,13 +690,13 @@ void SConfig::LoadDSPSettings(IniFile& ini)
 #elif defined __APPLE__
 	dsp->Get("Backend", &sBackend, BACKEND_COREAUDIO);
 #elif defined _WIN32
-	dsp->Get("Backend", &sBackend, BACKEND_XAUDIO2);
+	dsp->Get("Backend", &sBackend, BACKEND_CUBEB);
 #elif defined ANDROID
 	dsp->Get("Backend", &sBackend, BACKEND_OPENSLES);
 #else
 	dsp->Get("Backend", &sBackend, BACKEND_NULLSOUND);
 #endif
-	dsp->Get("Volume", &m_Volume, 100);
+	dsp->Get("Volume", &m_Volume, 25);
 	dsp->Get("CaptureLog", &m_DSPCaptureLog, false);
 
 	// fix 5.8b style setting
@@ -724,7 +737,7 @@ void SConfig::LoadAnalyticsSettings(IniFile& ini)
 
 	analytics->Get("ID", &m_analytics_id, "");
 	analytics->Get("Enabled", &m_analytics_enabled, false);
-	analytics->Get("PermissionAsked", &m_analytics_permission_asked, false);
+	analytics->Get("PermissionAsked", &m_analytics_permission_asked, true);
 }
 
 void SConfig::LoadBluetoothPassthroughSettings(IniFile& ini)
@@ -779,8 +792,12 @@ void SConfig::LoadDefaults()
 #endif
 
 	iCPUCore = PowerPC::CORE_JIT64;
-	iTimingVariance = 40;
+	iTimingVariance = 8;
+#ifdef IS_PLAYBACK
 	bCPUThread = false;
+#else
+	bCPUThread = true;
+#endif
 	bSyncGPUOnSkipIdleHack = true;
 	bRunCompareServer = false;
 	bDSPHLE = true;
@@ -792,10 +809,10 @@ void SConfig::LoadDefaults()
 	iBBDumpPort = -1;
 	iVideoRate = 8;
 	bHalfAudioRate = false;
-	iPollingMethod = POLLING_CONSOLE;
+	iPollingMethod = POLLING_ONSIREAD;
 	bSyncGPU = false;
 	bFastDiscSpeed = false;
-	m_strWiiSDCardPath = File::GetUserPath(F_WIISDCARD_IDX);
+	m_strWiiSDCardPath = "";
 	bEnableMemcardSdWriting = true;
 	SelectedLanguage = 0;
 	bOverrideGCLanguage = false;
@@ -807,8 +824,8 @@ void SConfig::LoadDefaults()
 
 	iPosX = INT_MIN;
 	iPosY = INT_MIN;
-	iWidth = -1;
-	iHeight = -1;
+	iWidth = 640;
+	iHeight = 430;
 
 	m_analytics_id = "";
 	m_analytics_enabled = false;
@@ -1132,19 +1149,28 @@ DiscIO::Language SConfig::GetCurrentLanguage(bool wii) const
 	return language;
 }
 
+// Hack to deal with 20XX images
+u16 SConfig::GetGameRevision() const { return m_revision; }
+std::string SConfig::GetGameID_Wrapper() const
+{
+	return m_gameType == GAMETYPE_MELEE_20XX ? "GALEXX" : GetGameID();
+}
+
+
+
 IniFile SConfig::LoadDefaultGameIni() const
 {
-	return LoadDefaultGameIni(m_gameType == GAMETYPE_MELEE_20XX ? "GALEXX" : GetGameID(), m_revision);
+	return LoadDefaultGameIni(GetGameID_Wrapper(), m_revision);
 }
 
 IniFile SConfig::LoadLocalGameIni() const
 {
-	return LoadLocalGameIni(m_gameType == GAMETYPE_MELEE_20XX ? "GALEXX" : GetGameID(), m_revision);
+	return LoadLocalGameIni(GetGameID_Wrapper(), m_revision);
 }
 
 IniFile SConfig::LoadGameIni() const
 {
-	return LoadGameIni(m_gameType == GAMETYPE_MELEE_20XX ? "GALEXX" : GetGameID(), m_revision);
+	return LoadGameIni(GetGameID_Wrapper(), m_revision);
 }
 
 IniFile SConfig::LoadDefaultGameIni(const std::string& id, u16 revision)
