@@ -16,7 +16,8 @@ class GeckoCode
 {
 public:
 
-	GeckoCode() : enabled(false)
+	GeckoCode()
+	  : enabled(false)
 	{}
 
 	struct Code
