@@ -66,10 +66,6 @@ SlippiNetplayClient::SlippiNetplayClient(const std::string &address, const u16 r
     , m_qos_flow_id(0)
 #endif
 {
-	// Initialize enet
-	auto res = enet_initialize();
-	INFO_LOG(SLIPPI_ONLINE, "Enet init res: %d", res);
-
 	WARN_LOG(SLIPPI_ONLINE, "Initializing Slippi Netplay for port: %d, with host: %s", localPort,
 	         isDecider ? "true" : "false");
 
@@ -233,7 +229,7 @@ unsigned int SlippiNetplayClient::OnData(sf::Packet &packet)
 	case NP_MSG_SLIPPI_MATCH_SELECTIONS:
 	{
 		auto s = readSelectionsFromPacket(packet);
-		ERROR_LOG(SLIPPI_ONLINE, "[Received Selections] Char: 0x%X, Color: 0x%X", s->characterId, s->characterId);
+		INFO_LOG(SLIPPI_ONLINE, "[Netplay] Received selections from opponent");
 		matchInfo.remotePlayerSelections.Merge(*s);
 
 		// Set player name is not empty
@@ -452,7 +448,15 @@ void SlippiNetplayClient::ThreadFunc()
 				enet_packet_destroy(netEvent.packet);
 				break;
 			case ENET_EVENT_TYPE_DISCONNECT:
-				m_do_loop.Clear(); // Stop the loop, will trigger a disconnect
+				ERROR_LOG(SLIPPI_ONLINE, "[Netplay] Disconnected Event detected: %s",
+				          netEvent.peer == m_server ? "same client" : "diff client");
+
+				// If the disconnect event doesn't come from the client we are actually listening to,
+				// it can be safely ignored
+				if (netEvent.peer == m_server)
+				{
+					m_do_loop.Clear(); // Stop the loop, will trigger a disconnect
+				}
 				break;
 			default:
 				break;
