@@ -8,6 +8,7 @@
 #include "Common/CommonPaths.h"
 #include "Common/FileUtil.h"
 #include "Common/Logging/Log.h"
+#include "Common/MsgHandler.h"
 #include "Common/StringUtil.h"
 #include "Common/Thread.h"
 
@@ -43,7 +44,7 @@ static void system_hidden(const char *cmd)
 }
 #endif
 
-static void RunSystemCommand(const std::string& command)
+static void RunSystemCommand(const std::string &command)
 {
 #ifdef _WIN32
 	_wsystem(UTF8ToUTF16(command).c_str());
@@ -52,12 +53,51 @@ static void RunSystemCommand(const std::string& command)
 #endif
 }
 
+static size_t receive(char *ptr, size_t size, size_t nmemb, void *rcvBuf)
+{
+	size_t len = size * nmemb;
+	INFO_LOG(SLIPPI_ONLINE, "[User] Received data: %d", len);
+
+	std::string *buf = (std::string *)rcvBuf;
+
+	buf->insert(buf->end(), ptr, ptr + len);
+
+	return len;
+}
+
+SlippiUser::SlippiUser()
+{
+	CURL *curl = curl_easy_init();
+	if (curl)
+	{
+		curl_easy_setopt(curl, CURLOPT_WRITEFUNCTION, &receive);
+		curl_easy_setopt(curl, CURLOPT_TIMEOUT_MS, 5000);
+
+		// Set up HTTP Headers
+		m_curlHeaderList = curl_slist_append(m_curlHeaderList, "Content-Type: application/json");
+		curl_easy_setopt(curl, CURLOPT_HTTPHEADER, m_curlHeaderList);
+
+#ifdef _WIN32
+		// ALPN support is enabled by default but requires Windows >= 8.1.
+		curl_easy_setopt(curl, CURLOPT_SSL_ENABLE_ALPN, false);
+#endif
+
+		m_curl = curl;
+	}
+}
+
 SlippiUser::~SlippiUser()
 {
 	// Wait for thread to terminate
 	runThread = false;
 	if (fileListenThread.joinable())
 		fileListenThread.join();
+
+	if (m_curl)
+	{
+		curl_slist_free_all(m_curlHeaderList);
+		curl_easy_cleanup(m_curl);
+	}
 }
 
 bool SlippiUser::AttemptLogin()
@@ -69,24 +109,18 @@ bool SlippiUser::AttemptLogin()
 	{
 		std::string userFilePathTxt =
 		    userFilePath + ".txt"; // Put the filename here in its own scope because we don't really need it elsewhere
-		// If both files exist we just log they exist and take no further action
-		if (File::Exists(userFilePathTxt) && File::Exists(userFilePath))
+		if (File::Exists(userFilePathTxt))
 		{
-			INFO_LOG(SLIPPI_ONLINE,
-			         "Found both .json.txt and .json file for user data. Using .json and ignoring the .json.txt");
-		}
-		// If only the .txt file exists copy the contents to a json file and delete the text file
-		else if (File::Exists(userFilePathTxt))
-		{
-			// Attempt to copy the txt file to the json file path. If it fails log a warning
-			if (!File::Copy(userFilePathTxt, userFilePath))
+			// If both files exist we just log they exist and take no further action
+			if (File::Exists(userFilePath))
 			{
-				WARN_LOG(SLIPPI_ONLINE, "Could not copy file %s to %s", userFilePathTxt.c_str(), userFilePath.c_str());
+				INFO_LOG(SLIPPI_ONLINE,
+				         "Found both .json.txt and .json file for user data. Using .json and ignoring the .json.txt");
 			}
-			// Attempt to delete the txt file. If it fails log an info because this isn't as critical
-			if (!File::Delete(userFilePathTxt))
+			// If only the .txt file exists move the contents to a json file and log if it fails
+			else if (!File::Rename(userFilePathTxt, userFilePath))
 			{
-				INFO_LOG(SLIPPI_ONLINE, "Failed to delete %s", userFilePathTxt.c_str());
+				WARN_LOG(SLIPPI_ONLINE, "Could not move file %s to %s", userFilePathTxt.c_str(), userFilePath.c_str());
 			}
 		}
 	}
@@ -100,6 +134,7 @@ bool SlippiUser::AttemptLogin()
 	isLoggedIn = !userInfo.uid.empty();
 	if (isLoggedIn)
 	{
+		overwriteFromServer();
 		WARN_LOG(SLIPPI_ONLINE, "Found user %s (%s)", userInfo.displayName.c_str(), userInfo.uid.c_str());
 	}
 
@@ -131,23 +166,29 @@ void SlippiUser::OpenLogInPage()
 	RunSystemCommand(command);
 }
 
-void SlippiUser::UpdateFile()
-{
-#ifdef _WIN32
-	std::string path = File::GetExeDirectory() + "/dolphin-slippi-tools.exe";
-	std::string command = path + " user-update";
-	system_hidden(command.c_str());
-#endif
-}
-
 void SlippiUser::UpdateApp()
 {
 #ifdef _WIN32
 	auto isoPath = SConfig::GetInstance().m_strFilename;
 
 	std::string path = File::GetExeDirectory() + "/dolphin-slippi-tools.exe";
-	std::string command = "start \"Updating Dolphin\" \"" + path + "\" app-update -launch -iso \"" + isoPath + "\"";
+	std::string echoMsg = "echo Starting update process. If nothing happen after a few "
+	                      "minutes, you may need to update manually from https://slippi.gg/netplay ...";
+	std::string command =
+	    "start /b cmd /c " + echoMsg + " && \"" + path + "\" app-update -launch -iso \"" + isoPath + "\"";
 	WARN_LOG(SLIPPI, "Executing app update command: %s", command);
+	RunSystemCommand(command);
+#elif defined(__APPLE__)
+#else
+	const char *appimage_path = getenv("APPIMAGE");
+	if (!appimage_path)
+	{
+		CriticalAlertT("Automatic updates are not available for non-AppImage Linux builds.");
+		return;
+	}
+	std::string path(appimage_path);
+	std::string command = "appimageupdatetool " + path;
+	WARN_LOG(SLIPPI, "Executing app update command: %s", command.c_str());
 	RunSystemCommand(command);
 #endif
 }
@@ -174,6 +215,11 @@ void SlippiUser::LogOut()
 	userInfo = emptyUser;
 }
 
+void SlippiUser::OverwriteLatestVersion(std::string version)
+{
+	userInfo.latestVersion = version;
+}
+
 SlippiUser::UserInfo SlippiUser::GetUserInfo()
 {
 	return userInfo;
@@ -198,19 +244,18 @@ void SlippiUser::FileListenThread()
 	}
 }
 
-// On Linux platforms, the user.json file lives in the Sys/ directory in
-// order to deal with the fact that we want the configuration for AppImage 
-// builds to be mutable. 
+// On Linux platforms, the user.json file lives in the XDG_CONFIG_HOME/SlippiOnline
+// directory in order to deal with the fact that we want the configuration for AppImage
+// builds to be mutable.
 std::string SlippiUser::getUserFilePath()
 {
 #if defined(__APPLE__)
-	std::string dirPath = File::GetBundleDirectory() + "/Contents/Resources";
+	std::string userFilePath = File::GetBundleDirectory() + "/Contents/Resources" + DIR_SEP + "user.json";
 #elif defined(_WIN32)
-	std::string dirPath = File::GetExeDirectory();
+	std::string userFilePath = File::GetExeDirectory() + DIR_SEP + "user.json";
 #else
-	std::string dirPath = File::GetSysDirectory();
+	std::string userFilePath = File::GetUserPath(F_USERJSON_IDX);
 #endif
-	std::string userFilePath = dirPath + DIR_SEP + "user.json";
 	return userFilePath;
 }
 
@@ -249,4 +294,38 @@ void SlippiUser::deleteFile()
 {
 	std::string userFilePath = getUserFilePath();
 	File::Delete(userFilePath);
+}
+
+void SlippiUser::overwriteFromServer()
+{
+	if (!m_curl)
+		return;
+
+	// Perform curl request
+	std::string resp;
+	curl_easy_setopt(m_curl, CURLOPT_URL, (URL_START + "/" + userInfo.uid).c_str());
+	curl_easy_setopt(m_curl, CURLOPT_WRITEDATA, &resp);
+	CURLcode res = curl_easy_perform(m_curl);
+
+	if (res != 0)
+	{
+		ERROR_LOG(SLIPPI, "[User] Error fetching user info from server, code: %d", res);
+		return;
+	}
+
+	long responseCode;
+	curl_easy_getinfo(m_curl, CURLINFO_RESPONSE_CODE, &responseCode);
+	if (responseCode != 200)
+	{
+		ERROR_LOG(SLIPPI, "[User] Server responded with non-success status: %d", responseCode);
+		return;
+	}
+
+	// Overwrite userInfo with data from server
+	auto r = json::parse(resp);
+	userInfo.connectCode = r.value("connectCode", userInfo.connectCode);
+	userInfo.latestVersion = r.value("latestVersion", userInfo.latestVersion);
+
+	// TODO: Once it's possible to change Display name from website, uncomment below
+	// userInfo.displayName = r.value("displayName", userInfo.displayName);
 }
