@@ -2,8 +2,11 @@
 
 #include "Common/CommonTypes.h"
 #include <atomic>
+#include <curl/curl.h>
 #include <string>
 #include <thread>
+#include <unordered_map>
+#include <vector>
 
 class SlippiUser
 {
@@ -18,14 +21,15 @@ class SlippiUser
 		std::string fileContents = "";
 	};
 
+	SlippiUser();
 	~SlippiUser();
 
 	bool AttemptLogin();
 	void OpenLogInPage();
-	void UpdateFile();
 	void UpdateApp();
 	void ListenForLogIn();
 	void LogOut();
+	void OverwriteLatestVersion(std::string version);
 	UserInfo GetUserInfo();
 	bool IsLoggedIn();
 	void FileListenThread();
@@ -34,9 +38,15 @@ class SlippiUser
 	std::string getUserFilePath();
 	UserInfo parseFile(std::string fileContents);
 	void deleteFile();
+	void overwriteFromServer();
 
 	UserInfo userInfo;
 	bool isLoggedIn = false;
+
+	const std::string URL_START = "https://users-rest-dot-slippi.uc.r.appspot.com/user";
+	CURL *m_curl = nullptr;
+	struct curl_slist *m_curlHeaderList = nullptr;
+	std::vector<char> receiveBuf;
 
 	std::thread fileListenThread;
 	std::atomic<bool> runThread;
