@@ -6,6 +6,7 @@
 #include <libusb.h>
 #include <mutex>
 #include <iostream>
+#include <chrono>
 
 #include "Common/Event.h"
 #include "Common/Flag.h"
@@ -47,11 +48,11 @@ static std::thread s_adapter_output_thread;
 static Common::Flag s_adapter_thread_running;
 
 static Common::Event s_rumble_data_available;
-static unsigned char s_latest_rumble_data[5];
 
 static std::mutex s_init_mutex;
 static std::thread s_adapter_detect_thread;
 static Common::Flag s_adapter_detect_thread_running;
+static Common::Event s_hotplug_event;
 
 static std::function<void(void)> s_detect_callback;
 
@@ -67,6 +68,9 @@ static u8 s_endpoint_out = 0;
 
 static u64 s_last_init = 0;
 
+static u64 s_consecutive_slow_transfers = 0;
+static double s_read_rate = 0.0;
+
 bool adapter_error = false;
 
 bool AdapterError()
@@ -74,20 +78,59 @@ bool AdapterError()
 	return adapter_error && s_adapter_thread_running.IsSet();
 }
 
+bool IsReadingAtReducedRate()
+{
+	return s_consecutive_slow_transfers > 80;
+}
+
+double ReadRate() 
+{
+	return s_read_rate;
+}
+
 static void Read()
 {
+	s_consecutive_slow_transfers = 0;
 	adapter_error = false;
+
+	u8 bkp_payload_swap[37];
+	int bkp_payload_size = 0;
+	bool has_prev_input = false;
+	s_read_rate = 0.0;
 
 	int payload_size = 0;
 	while (s_adapter_thread_running.IsSet())
 	{
-		int err = libusb_interrupt_transfer(s_handle, s_endpoint_in, s_controller_payload_swap,
-		                                    sizeof(s_controller_payload_swap), &payload_size, 16);
-		adapter_error = err != LIBUSB_SUCCESS && SConfig::GetInstance().bAdapterWarning;
+		bool reuseOldInputsEnabled = SConfig::GetInstance().bAdapterWarning;
+		std::chrono::time_point<std::chrono::high_resolution_clock> start = std::chrono::high_resolution_clock::now();
+		adapter_error = libusb_interrupt_transfer(s_handle, s_endpoint_in, s_controller_payload_swap,
+			sizeof(s_controller_payload_swap), &payload_size, 32) != LIBUSB_SUCCESS && reuseOldInputsEnabled;
 
-		if (err)
-			ERROR_LOG(SERIALINTERFACE, "adapter libusb read failed: err=%s", libusb_error_name(err));
+		double elapsed = std::chrono::duration_cast<std::chrono::nanoseconds>(std::chrono::high_resolution_clock::now() - start).count() / 1000000.0;
 
+		// Store previous input and restore in the case of an adapter error
+		if (reuseOldInputsEnabled)
+		{
+			if (!adapter_error)
+			{
+				memcpy(bkp_payload_swap, s_controller_payload_swap, 37);
+				bkp_payload_size = payload_size;
+				has_prev_input = true;
+			}
+			else if (has_prev_input)
+			{
+				memcpy(s_controller_payload_swap, bkp_payload_swap, 37);
+				payload_size = bkp_payload_size;
+			}
+		}
+
+		if(elapsed > 15.0)
+			s_consecutive_slow_transfers++;
+		else
+			s_consecutive_slow_transfers = 0;
+
+		s_read_rate = elapsed;
+		
 		{
 			std::lock_guard<std::mutex> lk(s_mutex);
 			std::swap(s_controller_payload_swap, s_controller_payload);
@@ -105,10 +148,9 @@ static void Write()
 	{
 		if (s_rumble_data_available.WaitFor(std::chrono::milliseconds(100)))
 		{
-			int err = libusb_interrupt_transfer(s_handle, s_endpoint_out, s_latest_rumble_data,
-			                                    sizeof(s_latest_rumble_data), &size, 16);
-			if (err)
-				ERROR_LOG(SERIALINTERFACE, "adapter libusb write failed: err=%s", libusb_error_name(err));
+			unsigned char rumble[5] = {0x11, s_controller_rumble[0], s_controller_rumble[1], s_controller_rumble[2],
+			                           s_controller_rumble[3]};
+			libusb_interrupt_transfer(s_handle, s_endpoint_out, rumble, sizeof(rumble), &size, 32);
 		}
 	}
 
@@ -121,11 +163,8 @@ static int HotplugCallback(libusb_context* ctx, libusb_device* dev, libusb_hotpl
 {
 	if (event == LIBUSB_HOTPLUG_EVENT_DEVICE_ARRIVED)
 	{
-		if (s_handle == nullptr && CheckDeviceAccess(dev))
-		{
-			std::lock_guard<std::mutex> lk(s_init_mutex);
-			AddGCAdapter(dev);
-		}
+		if (s_handle == nullptr)
+			s_hotplug_event.Set();
 	}
 	else if (event == LIBUSB_HOTPLUG_EVENT_DEVICE_LEFT)
 	{
@@ -158,22 +197,16 @@ static void ScanThreadFunc()
 
 	while (s_adapter_detect_thread_running.IsSet())
 	{
+		if (s_handle == nullptr)
+		{
+			std::lock_guard<std::mutex> lk(s_init_mutex);
+			Setup();
+		}
+
 		if (s_libusb_hotplug_enabled)
-		{
-			static timeval tv = { 0, 500000 };
-			libusb_handle_events_timeout(s_libusb_context, &tv);
-		}
+			s_hotplug_event.Wait();
 		else
-		{
-			if (s_handle == nullptr)
-			{
-				std::lock_guard<std::mutex> lk(s_init_mutex);
-				Setup();
-				if (s_detected && s_detect_callback != nullptr)
-					s_detect_callback();
-			}
 			Common::SleepCurrentThread(500);
-		}
 	}
 	NOTICE_LOG(SERIALINTERFACE, "GC Adapter scanning thread stopped");
 }
@@ -226,6 +259,7 @@ void StopScanThread()
 {
 	if (s_adapter_detect_thread_running.TestAndClear())
 	{
+		s_hotplug_event.Set();
 		s_adapter_detect_thread.join();
 	}
 }
@@ -308,6 +342,12 @@ static bool CheckDeviceAccess(libusb_device* device)
 				ERROR_LOG(SERIALINTERFACE, "libusb_detach_kernel_driver failed with error: %d", ret);
 			}
 		}
+		// This call makes Nyko-brand (and perhaps other) adapters work.
+		// However it returns LIBUSB_ERROR_PIPE with Mayflash adapters.
+		const int transfer = libusb_control_transfer(s_handle, 0x21, 11, 0x0001, 0, nullptr, 0, 1000);
+		if (transfer < 0)
+			WARN_LOG(SERIALINTERFACE, "libusb_control_transfer failed with error: %d", transfer);
+
 		// this split is needed so that we don't avoid claiming the interface when
 		// detaching the kernel driver is successful
 		if (ret != 0 && ret != LIBUSB_ERROR_NOT_SUPPORTED)
@@ -349,11 +389,11 @@ static void AddGCAdapter(libusb_device* device)
 
 	int tmp = 0;
 	unsigned char payload = 0x13;
-	libusb_interrupt_transfer(s_handle, s_endpoint_out, &payload, sizeof(payload), &tmp, 16);
+	libusb_interrupt_transfer(s_handle, s_endpoint_out, &payload, sizeof(payload), &tmp, 32);
 
 	s_adapter_thread_running.Set(true);
-    s_adapter_input_thread = std::thread(Read);
-    s_adapter_output_thread = std::thread(Write);
+	s_adapter_input_thread = std::thread(Read);
+	s_adapter_output_thread = std::thread(Write);
 
 	s_detected = true;
 	if (s_detect_callback != nullptr)
@@ -417,16 +457,6 @@ GCPadStatus Input(int chan)
 	if (s_handle == nullptr || !s_detected)
 		return{};
 
-	if(AdapterError())
-	{
-		GCPadStatus centered_status = {0};
-		centered_status.stickX = centered_status.stickY =
-		centered_status.substickX = centered_status.substickY =
-		/* these are all the same */ GCPadStatus::MAIN_STICK_CENTER_X;
-
-		return centered_status;
-	}
-
 	int payload_size = 0;
 	u8 controller_payload_copy[37];
 
@@ -441,9 +471,9 @@ GCPadStatus Input(int chan)
 	if (payload_size != sizeof(controller_payload_copy) ||
 		controller_payload_copy[0] != LIBUSB_DT_HID)
 	{
+		// This can occur for a few frames on initialization.
 		ERROR_LOG(SERIALINTERFACE, "error reading payload (size: %d, type: %02x)", payload_size,
 			controller_payload_copy[0]);
-		Reset();
 	}
 	else
 	{
@@ -547,10 +577,6 @@ static void ResetRumbleLockNeeded()
 
 	std::fill(std::begin(s_controller_rumble), std::end(s_controller_rumble), 0);
 
-	s_latest_rumble_data[0] = 0x11;	
-	for (int i = 0; i < 4; i++)
-		s_latest_rumble_data[i + 1] = s_controller_rumble[i];
-		
 	s_rumble_data_available.Set();
 }
 
@@ -564,18 +590,7 @@ void Output(int chan, u8 rumble_command)
 		s_controller_type[chan] != ControllerTypes::CONTROLLER_WIRELESS)
 	{
 		s_controller_rumble[chan] = rumble_command;
-
-		unsigned char rumble[5] = { 0x11, s_controller_rumble[0], s_controller_rumble[1],
-															 s_controller_rumble[2], s_controller_rumble[3] };
-		int size = 0;
-
-		libusb_interrupt_transfer(s_handle, s_endpoint_out, rumble, sizeof(rumble), &size, 16);
-		// Netplay sends invalid data which results in size = 0x00.  Ignore it.
-		if (size != 0x05 && size != 0x00)
-		{
-			ERROR_LOG(SERIALINTERFACE, "error writing rumble (size: %d)", size);
-			Reset();
-		}
+		s_rumble_data_available.Set();
 	}
 }
 
